WIP music functions

+++ app.js
@@ -9,6 +9,8 @@ const client = new Discord.Client();
 // Here we load the config.json file that contains our token and our prefix values.
 const config = require("./config.json");
 const sql = require("sqlite");
+const ytdl = require('ytdl-core');
+const queue = new Map();
 sql.open("./scores.sqlite");
 // config.token contains the bot's token
 // config.prefix contains the message prefix.
@@ -61,6 +63,8 @@ client.on("message", async message => {
   // args = ["Is", "this", "the", "real", "life?"]
   const args = message.content.slice(config.prefix.length).trim().split(/ +/g);
   const command = args.shift().toLowerCase();
+  
+  const serverQueue = queue.get(message.guild.id);
 
   sql.get(`SELECT * FROM scores WHERE userId ="${message.author.id}"`).then(row => {
     if (!row) {
@@ -74,12 +78,107 @@ client.on("message", async message => {
     });
   });
 
+async function execute(message, serverQueue){
+	const args = message.content.split(' ');
+	const voiceChannel = message.member.voiceChannel;
+
+	if(!voiceChannel) return message.channel.send('You need to be in a voice channel.');
+	const permissions = voiceChannel.permissionsFor(message.client.user);
+	if(!permissions.has('CONNECT') || !permissions.has('SPEAK')){
+	   return message.channel.send('I need permissions to speak and join your channel.');
+	   }
+
+	const songInfo = await ytdl.getInfo(args[1]);
+	const song = {
+		title: songInfo.title,
+		url: songInfo.video_url,
+		};
+	
+	if (!serverQueue){
+		const queueConstruct = {
+			textChannel: message.channel,
+			voiceChannel: voiceChannel,
+			connection: null,
+			songs: [],
+			volume: 5,
+			playing: true,
+		};
+
+		queue.set(message.guild.id, queueConstruct);
+		queueConstruct.songs.push(song);
+
+		try {
+			var connection = await voiceChannel.join();
+			queueConstruct.connection = connection;
+			play(message.guild, queueConstruct.songs[0]);
+		} catch (err) {
+			console.log(err);
+			queue.delete(message.guild.id);
+			return message.channel.send(err);
+			}
+		} else {
+			serverQueue.songs.push(song);
+			console.log(serverQueue.songs);
+			return message.channel.send('${song.title} has been added to the queue!');
+		}
+
+	}
+
+  function skip(message, serverQueue) {
+	if (!message.member.voiceChannel) return message.channel.send ('You have to be in a voice channel.');
+	if (!serverQueue) return message.channel.send('There is no song to skip!');
+	serverQueue.connection.dispatcher.end();
+
+  }
+
+  function stop(message, serverQueue) {
+	if (!message.member.voiceChannel) return message.channel.send('You have to be in a voice channel to stop.');
+	serverQueue.songs = [];
+	serverQueue.connection.dispatcher.end();
+  }
+
+  function play(guild, song) {
+	const serverQueue = queue.get(guild.id);
+
+	if (!song) {
+		serverQueue.voiceChannel.leave();
+		queue.delete(guild.id);
+		return;
+	}
+
+	const dispatcher = serverQueue.connection.playStream(ytdl(song.url))
+		.on('end', () => {
+			console.log('Music ended.');
+			serverQueue.songs.shift();
+			play(guild, serverQueue.songs[0]);
+		})
+		.on('error', error => {
+			console.error(error);
+		});
+	dispatcher.setVolumeLogarithmic(serverQueue.volume / 5);
+  }
+
+  if(command === "play"){
+	execute(message, serverQueue);
+	return;
+	}
+  if(command === "skip"){
+	skip(message, serverQueue);
+	return;
+	}
+  if(command === "stop"){
+	stop(message, serverQueue);
+	return;	
+	}
+
   if(command === "say") {
     const sayMessage = args.join(" ");
     message.delete().catch(O_o=>{});
     message.channel.send(sayMessage,{tts: true});
     return;
   }
+
+
 
    if (command === "rps"){
